+++ bless/backends/bluezdbus/dbus/advertisement.py
